Ruslan is simply right, for these preperiodic guys there is no shift, we do not need skype meeting for  that (but we do need lot of meetings the from the day this is on arXiv). Please finalize the relevant text: there are
1) full-space RPOs (intrinsic continuous shift, no reflection)
2) full-space pre periodic  RPOs - or whatever professionals call these things - (reflection, no intrinsic shift)
3) symmetric subspaces POs, no shifts

What makes me uneasy is the high frequency of these solutions - Ruslan, can you run a long search for them for up to T_p=200? If computing power is the problem, Vaggelis can run it on the cluster we have access to. I would like to see whether they are as frequent as the card-carying RPOs, and have similar unstable eigs/Lyapunovs and similar nos of unstable ones - that would be plotted color-coded in fig 6.1(b), caption for 6.1(a) should go back to the original Ruslan's text.

Wonder how this will show up in symbolic dynamics....




On 9/14/07, Predrag Cvitanovic <predrag.cvitanovic@physics.gatech.edu> wrote:

    Guys

-- please check my summary of deliberations so far
-- please edit the Summary section to its final form
-- if possible, compute many pre-periodic orbits up to
   t$T_p < 200$.
-------------------------
    I agree with Ruslan - R\tau_l generates a family that is preperiodic to $( R\tau_l)^2=1$. So
    it is like $( R)^2=1$ described in ChaosBook, and maybe that is all that is to it. Still,
    we need to think  bit - please read up on irreps of O(2). Unlike SO(2), reflection/inversion
    makes this a nonabelian group and I think that is what is confusing us.  Many people have
    looked equivariant POs before us...