Changed:
( We mention D_m, but only focus on D_1, D_2 and D_3 because D_1
(or D_2, depending on ones perspective) is relevant 
to the RPOs with reflection, and ... )

Added:

(Added references to Krauskopf et al, Johnson et al)

I think we are ready for submission! SIADS only requires pdf
file at this point but I will copy all source code in SIADS-v2.
Then I will go for arxiv-v2 this afternoon.