Changed discussion of symmetry of POs to avoid complications and mistakes. In response2.1 changed:
( We mention D_m, but only focus on D_1, D_2 and D_3 because
all equilibria are in the antisymmetric subspace (the fixed-point
subspace of D_1), while E_1, E_2, E_3 have symmetry D_1, D_2,
D_3 respectively. D_1 is also relevant 
to the RPOs with reflection. General case...

Finalized text is now in SIADS-v2.