Added some figures. Let me know if there is anything I am forgetting. All of them but the RedTraj ones are up to a few minor modifications, essentially finished.
I put in multiple for the reduced trajectory. 3 of them are just different views with a plane passing through the near singularity points. The plane was chosen to be normal to the plane the attractor is in. The other two (the noplane ones) are two different views with different boxes put around the part of the trajectory used for the others.

Let me know if you want the plane and what view of the reduced trajectory looks better.

I can show up tomorrow, but I'll be gone from the 22nd to the 26th.