B&W figures with E1 and TW1 labels added.
I can't seem to find any slices that will work.
The singularities seem to occur at nearly the same time for all the slices - whenever the full space trajectory passes near the z-axis (which it does regularly).
It seems like the trajectory can pass by the z-axis in any arbitrary direction, so I can't think of anyway to choose two slices so that it will never have a singularity in both
at the same time.