To Predrag and Ruslan: I am very glad you two got again into a symmetry reduction argument, and I will try to interupt it by proposing a third way (to keep you both happy or unhappy). I needed to revisit the problem because Predrag is correct - symmetry reduction is needed in what Kazz and Hugues do (and in what they want me to do).
My proposition is based on rewriting (with pen and paper) invariant
variables obtained from slicing into a polar form. The second step is to
invoke L. Tuckerman's insight on polar variables singularities to modify
the invariants to avoid slice-induced jumps. Taking care of reflections
is the final step. Since there is no computer algebra used, the final
variables have a simple form which can be easily used in numerics - in
any dimension and with no extra man-power. 

It is all in ksReduced directory, as a single section in a paper which I propose
the three of us write. I meant to write a real first draft before I upload it, but
it seems you are both into the subject again, and in any case I will
have to explain very soon to Kazz and Hugues how I implement automated
searches for nearby orbits.

I suggest comments go into siminos/blog, to keep siminos/ksReduced as
article-like as possible.