Is there an equilibrium missing?
Burak's result is in agreement with expectations from degree of
multivariate polynomials but there's a solution that we know
analytically that is missing.

Also half assed answered Predrag... not so sure what "Lyapunov exponent"
code is doing anymore now that I re-read the Wolf paper. Will stop by
tomorrow to discuss.