written some introductory parts of the article
I've written a few pages of the article, but I think it is a very bad
idea and bad PhD training that a senior adviser writes articles.
Students in Spring 2012 chaos class pulled that trick, but I was not
their formal adviser, I had no choice - this time, please write the
article, and I'll edit or suggest modifications to the existing text. Or
blog your argument why you should be a co-author exempted from writing
:)