an error in the multishooting formula?
I started rewriting the multiple shooting method for search for the \rpo
s, with intention of including it into ChaosBook. As an example?
exercise? in an appendix? - in any case have to include the calculation
for a \reqv, possibly in the Lagrange multiplier formulation, if that is
easier to understand...

I stopped, however, until Burak and I agree on my correction to his
initial formula.