codes for hexagonal lattice mapping
will upload descriptions later in the tex file.

For Pavel, the linkminsearch function will find a physical solution
providing a symbolic sequence. The details of this method is described
in Schreiber's paper.

All the functions under the hexfunc folder can compute the hexagonal
mapping in the fundamental domain for any type of scattering potential
(soft potentials, etc). They are created for later study of soft
potential Lorentz gas.