a different elementary cell
Perhaps worth a thought: Hasselblatt and define
the `fundamental domain' (our elementary cell) of the translation group
as a parallelepiped one triangle-high, three triangles wide, rather than
our hexagon. Why? instead of identifying three pairs of opposite edges
of a hexagon, one identifies two pairs of opposite edges of a
parallelepiped, thus quotiented dynamics is on a 2-torus, just as for
tiling with squares.