Major group theory developments
1) working through Alcock-Zeilinger & Weigert #1. Looks good. No wonder
Henriette did not do it (or I alone) - it is a PhD length project
2) Julia has discovered my E_7 family in a table of "Exotic Symplectic
Holonomies". Will have to contact the authors. Pain expected.
3) Shimizu and Tachikawa credit Deligne for "Deligne’s exceptional
series of groups", also independently found by Cvitanovic. I do not get
the ethics of mathematicians. If I discover the series, and Deligne
"rediscovers" it 20 years later, fully aware that I had discovered it,
how is that "Deligne series"?